moving annotations
