+++ Wire.PerfTest/Wire.PerfTest.csproj
@@ -73,8 +73,8 @@
       <Private>True</Private>
     </Reference>
     <Reference Include="MessageShark, Version=1.0.1.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\MessageShark.1.0.2\lib\net40\MessageShark.dll</HintPath>
-      <Private>True</Private>
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\MessageShark\src\MessageShark\MessageShark\bin\Release\MessageShark.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.CodeAnalysis, Version=2.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\packages\Microsoft.CodeAnalysis.Common.2.0.0-beta3\lib\net45\Microsoft.CodeAnalysis.dll</HintPath>
