+++ Wire/project.json
@@ -1,5 +1,5 @@
 ﻿{
-  "version": "0.8.0-*",
+  "version": "0.8.1-*",
   "description": "Wire, fast binary POCO serializer",
   "authors": [
     "Roger Johansson"
